fixed bugs and refactored navigation stack
