+++ .env.local
@@ -1,0 +1,1 @@
+REACT_EDITOR=code
