+++ navigations/settingStack.js
@@ -1,0 +1,30 @@
+import 'react-native-gesture-handler';
+import { createStackNavigator } from '@react-navigation/stack';
+import { NavigationContainer} from '@react-navigation/native';
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import {Image} from 'react-native-elements';
+import { MaterialCommunityIcons } from '@expo/vector-icons';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import Dashboard from '../screens/Dashboard';
+import CreateGroup from '../screens/CreateGroup';
+
+const Stack = createStackNavigator();
+
+export function SettingsStack() {
+  return (
+    <Stack.Navigator
+      initialRouteName="Settings"
+      screenOptions={{
+        headerStyle: { backgroundColor: '#3DD5F4' },
+        headerTintColor: '#fff',
+        headerTitleStyle: { fontWeight: 'bold' },
+      }}>
+      <Stack.Screen
+        name="Settings"
+        component={SettingsScreen}
+        options={{ title: 'Setting Page' }}
+      />
+    </Stack.Navigator>
+  );
+}
